requirements.txt
openai==0.28.1
PyQt5==5.15.9
pandas==2.0.3
requests==2.31.0
python-dotenv==1.0.1
pyyaml==6.0.1
psutil==5.9.8  # For system information
platformdirs==4.2.0  # For platform-specific paths 